Add files via upload
My Mabe project is my prof of being able to try things one step at a time, building up small successes along the way. After I understood the assignment better, I started making amll edits creating files and splitting what needed to e split and creating the ones that needed to be created from scratch. I for some reason had errors everywhere that I tried to fix but I think it was because if my compiler or the routing of things in my computer. I then didn't know how to connect the files to the old file so I went to the professor and asked about how I can do that. They showed me how to d it but unfortunately I didn't have enough time to fix it up as I got caught up with finals, I am planing on making it work this summer when I have more free time, but i am also almost certain that as soon as I connect it the code will definitely run.

I also only turned in collable because the Mabe file is too big to submit